Storepage refactored, new stores features added separated
