+++ src/features/store/components/StoreSearchBar/StoreSearchBar.css
@@ -1,0 +1,15 @@
+.storeSearchBar {
+    @apply mb-6 flex flex-wrap items-end gap-3;
+}
+
+.storeSearchBarName {
+    @apply mb-0 min-w-[220px] flex-1;
+}
+
+.storeSearchBarStatus {
+    @apply mb-0 min-w-[180px];
+}
+
+.storeSearchBarAction {
+    @apply mb-0;
+}
