+++ src/features/store/components/StoreSearchBar/index.ts
@@ -1,0 +1,1 @@
+export * from './StoreSearchBar';
