+++ src/pages/store/StoresPage.css
@@ -1,0 +1,15 @@
+.storesPage {
+    @apply w-full;
+}
+
+.storesPageHeader {
+    @apply mb-6;
+}
+
+.storesPageTitle {
+    @apply text-2xl font-bold;
+}
+
+.storesPageAlert {
+    @apply mb-4 py-2 text-sm;
+}
